ADT: Refactoring classes dealing with android resources out of the editor.
Basically:
editors.resources.manager -> resources.manager
editors.resources.configurations -> resources.configurations

This is to make it less confusing between the "Resources editors" and the
class parsing/handling Android resources (either in a project or in the
framework).
Also moved the ResourceExplorerView out of the resources editors, and clean
up a few other misc classes.