SDK #1800921: Packages the Windows SDK in a temporary location.
This changes make_windows_sdk so that it packages everything
in a temp dir. Once the build & packaging succeeds, the final
Zip is copied to the desired location.

The temp dir is *not* cleared on zip errors, which helps in
diagnosing packaging issues after the fact.

This has no impact on the build process itself (i.e. make
adb, emulator, etc.) which all happen in $SDK/out anyway.

The temp dir can be passed as an argument. If not provided,
the default Cygwin $TMP is used. Otherwise /tmp is used.

I also added an usage description when parameters are missing.