SDK Updater: provide a "force http" setting.
Note that the setting operates on every source by converting
the https:// links into http:// before fetching the sources
or the archives.

This also reorganize some things: the RepoSourceAdapter now
takes the UpdaterData as parameter (so both Adapters work the
same) and the label/contentProviders are nested classes instead
of static classes. Cleanup some stuff that was no longer useful
such as the RepoSources.TaskFactory (it's in UpdaterData already).

With this change I can install from the dl site over http at home!