Fix NPE that prevented any package w/o license to be installed.
java.util.Properties#setProperty() doesn't like it when the value is null
which was the case when a package had no license.

Made sure it won't happen on other properties than the license as well.

Also improved error display when an unexpected Throwable is thrown during
install (Stack Call is now display if the Throwable has no message).

BUG:2037085