Complete AdbWinApi split: submit prebuild binaries.
As the result of the split, now we have two DLLs that deal with USB driver
connection on Windows. AdbWinApi.dll deals with legacy driver and LoadLibrary
AdbWinUsbApi.dll on condition that WINUSB.DLL is installed on user's machine.
When both DLLs must be installed in the same location in SDK.
BUG 2033924