SDK Manager: display packages revisions
This changes the short description of all the package types
to display the revision.

It also changes the long description to be "whatever" is in the
XML; if there's nothing in the xml, use the short description.
Also appends the requirements for addon/platform and min-tools-rev.

I will address the packaging script to put proper descriptions
when generating the XML.

SDK BUG 2136068

Change-Id: Iaca7692f048f7f63111c2fe1f25c5588f0fc2099