Fix the locale selection in the config selector.
default locale means nothing. Instead we replace it with
- "any" if there are no resources specific to a locale in the project.
- "other" if there are resources specific to a locale in the project.

In both case, the actual locale used by the config is hard-coded to
a value that cannot be used by the resources ("__" for the language,
and "__" for the region), to enforce using the default resources.
Using a config with no language nor region had the effect to take
the first resources with a valid qualifier for language and/or
region.

Change-Id: I67a34d1a026a5d3d5e1438d3763ae27c5ed0ce55