ADT/GLE: Fix issue with the locale list refresh in the ConfigComposite
One of the problem stemmed from the ResourceMonitor sending removed
resource events to the ResourceManager first and then to other listeners.
Other listeners then failed to get a corresponding ResourceFolder
or ResourceFile since it was deleted by the ResourceManager.

Solution: add a resource-specific change listener in the ResourceManager.
Using this listener is much more efficient as the ResourceFolder/File
do not need to be figured out by the listener from the IFolder/IFile.

Also renamed the ResourceMonitor as it's confusing (it listens to
Eclipse resource changes, not Android project resource changes).
Since it deals mostly with projects and their resources, and is
global and used by other monitor, it's now called GlobalProjectMonitor.

Also cleaned up the ResourceManager and LayoutReloadMonitor to
implements their interface as internal classes so that the
methods do not show up in their public API.

Change-Id: I1967a48364fd946c98d9eee5bd4a333fc01bea2d