ADT/GLE: Fix config selector to not select a config that has a better match than the current file.
When replacing a file with another one (because of a user "open action")
it is possible the config selector will find the current config compatible,
even though the previous layout was a better for match for it.

The config selector now attemps to find a config for which the new
file is the best match.

Change-Id: I7d794c2a8b9a90a120970049cb402f9ee84f8749