ADT/GLE: Dynamically create ID values for non saved layout
When a non saved layout is rendered it is possible that some
views reference IDs that don't yet exists in the R class
because those IDs are created on the fly through @+id/...

For some layouts (RelativeLayout) this has a huge impact
on the rendering to the point of making useless. By
dynamically creating unknown ID, we can have a correct
rendering even if the layout is non saved.

Bug: 2370047
Change-Id: I09a0dfaa9d2a767229ab778438256a7609d6d337