Strawman support for streaming method profiling. do not merge.
This will send MPSS/MPSE packets instead of MPRS/MPRE when the client VM
supports them.  At least, it would if the property check weren't
hard-wired to false, which is done because the MPSE handling in DDMS is
incomplete (currently just dumps the first 100 chars so you can see that
it really got something).

Integrated from master to be in SDK Tools r5.

For bug 2160407.

Change-Id: I92b7d43fa78e3cb3eefdcef9e7357bbc6818822b