Strawman support for streaming method profiling.
This will send MPSS/MPSE packets instead of MPRS/MPRE when the client VM
supports them.  At least, it would if the property check weren't
hard-wired to false, which is done because the MPSE handling in DDMS is
incomplete (currently just dumps the first 100 chars so you can see that
it really got something).

For bug 2160407.