Strawman support for streaming hprof dumps. Do not merge
Send an HPDS instead of HPDU when the client VM supports it.  Once
additional support is added to DDMS, this will allow grabbing an hprof
dump through DDMS even if the target app doesn't have permission to write
to /sdcard.  (The down side is that the dump must be held on the native
heap while it's generated, temporarily bloating the process by a few
megabytes.)

Integrated from master for inclusion in the SDK Tools r5

For bug 2092855.

Change-Id: I54492c62cce3c52b0ac523f6e87b9438d2ab92c0