Fix https download for the standalone version of the SDK Updater
The problem was due to the launch script completely replacing
java.ext.dirs preventing the VM from accessing the optional
jce.jar package (Java Cryptography Extension).

The launch script now queries the JVM for the current value
of java.ext.dirs (using archquery) to append to it instead
of replacing it.

In order to do that, archquery.jar now outputs only the value
of the queried system property instead of output "name=value"

Bug: 2248892
Change-Id: If17d48e46529bebc1a25daf99f2748e9ec81a214