GLE2: IViewRule callback for selection.
Details:
- This moves the selection drawing from the LayoutCanvas to
  the groovy script.

- There's a single IViewRule "onSelected" method. The script
  which implement it return a closure (aka an anonymous method)
  that does the actual drawing.

- onSelected is called during the actualy selection or when
  the canvas is refreshed. The rendering closure is actually called
  later during the canvas onPaint.

Here I am mostly testing the logic of having a script callback
handle an event (the selection) and returning a closure that
does the drawing later (during onpaint).
The main rationale for separating them is that the GC needed to
draw isn't available during the actual selection event.

I plan to reuse this mechanism for drag'n'drop & co.

TODO: layout selection is broken and will be fixed in another CL.
TODO: onChildSelected will be next.
TODO: Originally I had plans for onSelected(isMultipleSelection),
      which I'm not sure I'll actually add later anyway., I fail to
      see the point in implementation.

Change-Id: Ibc5fecbd107429df1f53b2b8172aba55e6289795