Clean up for how the ADT plug-in keeps track of project and their targets.
For some the (IProject, IAndroidTarget) map was only filled when a project
was created or had its target changed. Other projects that were just opened
in the workspace wouldn't be put in the map, making each call to Sdk.getTarget()
read the default.properties.

This commit adds a new class ProjectState to centralize the state of a project
including target info, apk settings and soon libraries.
The project state are stored statically in Sdk so that they survive Sdk reload.

I'm cleaning up the API of Sdk also to make it let complex and weird.

Change-Id: I1324c622a646486a474ce26b2172385f00091b3f