ADT: Library support: source folder and pre-compiler.
This is the first step in the library support.
For each library, create a source folder in the main project that
is linked to the source folder of the library project.

The linked resources use a path variable named after the library
in the format: _android_<library name>.
These variables are always created when the link is created.

For now the link is recreated all the time, but we could
do a check and not redo it if it's already done.

Additionally, the pre-compiler creates the R class from
the res folders of the main and library projects.

Some misc fixes/clean-ups:
* Fix an issue with the new ProjectState where opening a
  project would not trigger a load of its target data.

* Changed the lock for all SDK operation:
  - moved the lock in Sdk accessible as Sdk.getLock()
  - made the few Sdk method that used their own synchronize
    block use the same lock as all others.

* removed the builders project and moved its content to sdklib
  This was meant as a way to share code between the Eclipse
  builders and the Ant tasks but sdklib is already used by
  both, so it's better to put the code in sdklib than
  have yet another project.

Change-Id: Ibfa449c7a809f28e428c03bbda8215969717ecde