Prevent AVD Manager in ADT from forcing reload of the targets.
When the AVD Manager launches, it would notify some listener
that the SDK is changed. This was because it used the SDK Changed
event as a way to initialize the pages.

In ADT, an extra listener actually reload the SDK manager
inside ADT when that event triggers, making the SDK also
reload when the AVD Manager window is opened.

The fix is to add a flag to the listener interface to
distinguish initial load and actual SDK change.

Change-Id: I395dc305fa0d698889cdbd7bb7db7d540dde568b