ADT: Fix NPE when SDK Manager uses a null ISdkLog.
SDK Bug: 2477460

Note: this fixes the NPE but it also makes sure we give
an ISdkLog even when calling from the toolbar SDK Manager
button or from the AVD Selector > Manager button.

Maybe it would be better to just assume the logger is never
null, remove the log==null checks all over the place and
correctly comment "log cannot be null" in the methods javadocs.

Change-Id: I50f65d698b7507b5e9eca866a332290f92edd05a