ADT: properly handle removed library project.
This edits the projects referencing the library to remove
the linked folder from the build path and delete it.

If the library is actually deleted this also remove the
path variable for the linked resource.

Also, force a recompile when a library project is opened
and linked to a main project. Apparently adding a new
source folder triggers only an incremental build instead
of a full (like removing a source folder does), and this
doesn't force recompiling the resources.

Change-Id: I005840bfa5c381160e5658755ebeec2db3eab19d