ADT: Fix library handling during project renaming.
When implmenting the full open/close support, I moved
libraryState.close() to the job when closing a library
project.

Since this is asynchronous, the library state wasn't
correct when projectOpen/Renamed were called, breaking
proper handling of a library project being renamed.

Moved back the call to close in the projectClosed
callback.

Change-Id: Iabf03a9c068bc6246bd162e46bc3111d8f52c434