Move ant.prop to sdk.prop and up a folder.
this file is meant to have general sdk properties, not
just Ant ones.

Also separated the library support from the Ant build version.
It's clearer whether or not libraries are supported, and it's
not just tied to Ant anyway (ADT needs to know as well).

Finally use that new property to do checks on library support
in both the custom Ant tasks and ADT.

Depends on a CL in development.git (sdk.atree change).

Change-Id: I77d229ed3fd60f0468b1d3d31f7cf147b03a32fb