Add support for using newer Ant rules distributed by the tools.
Right now, the Ant rules are packaged with each platform. This
is a problem when we want to release newer versions of the rules,
as we would have to respin newer rev for all major platforms just
for that.

The new mechanism has the Ant rules both in the tools folder
and in the platforms. The platform indicates which versions
it supports (support is based on the features of external tools
such as aapt, aidl and dex). The custom tasks can
figure out a newer but still compatible (ie not relying on
newer options in aapt/aidl/dex) version from the tools folder
and then uses this newer version.

Change-Id: I763533ec85fa9647ecf96ac6cbaefbfd1ad68fb0