Detect conflict due to change in number/list of ABI being exported.
A change in ABI (number, or one replacing another) in the ABIs of a project
would trigger a change in the generated build info, so we force
increment of versionCode in this case.

Also now display the actual project property that changed and is
causing a conflict for better troubleshooting.

Change-Id: Ibb654022f1bed0e81fa0c6e649dc69dd98f2f9e0