Stop using java.ext.dirs to modify classpath for android.bat
If the original value of java.ext.dirs contains space, then the
value is lost and we lose access to optional crypto packages.

The new mechanism uses a mix of -classpath (for main package
and swt whose path is computed automatically). For other
dependencies, we use the Class-Path attribute in the manifest
of sdkmanager.jar.

Change-Id: I06c46eb1abcd9b5bab6360d1430c0dcbc97f73ce