ADT GLE2: Access AttributeInfo from groovy scripts.
This CL extracts IAttributeInfo from DeclareStyleable -- this used
to be the metadata extracted from an attribute declared in attrs.xml.

The interface that is now available to groovy scripts lets them
check whether an attribute is a reference, a string, etc. For an
enum or flag that means having access to the defined values too.

In the case of the AbsoluteLayout drop, we use this info to check if
an attribute is a reference. For example when duplicating a
RelativeLayout dropped in an AbsoluteLayout, we need to create new
ids and the RelativeLayout is prone to have inner self-references
that can be automatically adjusted that way.

A large part of the CL is about refactoring the various manually-created
attributes in the Descriptors for resources, XML, and such. In this
case the IAttributeInfo that is made available via the AttributeDescriptor
is not used, but it made sense to keep this consistent in case we want
to expose it later and to avoid having to care about special cases.

Change-Id: Ic02f69bc16f0e549fcac3956ae93b86651c5be3c