Remove hard-coded source folder name for library projects in ADT.
Previously, the library mechanism expected the library to have a single source folder called 'src'.

There can now be more than one source folder (but the 'gen' folder of the library is always excluded),
and they don't need to be directly under the project folder. For example src/java is supported.

Change-Id: If139bd69cb720c331a1e5d1543a0cd4a91b8d675