ADT: Fix NPE in ConfigComposite
For some reason the locale used by the ConfigState in the
ConfigurationComposite had null members. Failed to reproduce
it. One clue might be there was a non-selected GLE2 open
as an editor tab but the Java workspace was not the main
visible one. It's possible some event fired before the
composite controls got created.

Change-Id: I2c8b6c86842f3c772c2a7ec113941eb8fc01b9e4