Allow for platform projects to be used with the "debug running app" button.
Since there's no way to figure out automatically which project is valid
for a given running app, we read the name of the project through an
environment variable.

NOTE: The goal of this new feature is for people working on the
platform but who still want to use the DDMS plugin and some of its
features.

Change-Id: I6f124c0413dab5c9f1fb240a3b311ba7234c4378