ADT GLE2: synchronized selection between canvas, outline and properties.
This CL adds a LayoutCanvasViewer, a JFace viewer wrapping the
LayoutCanvas control. This allows the canvas to participate in the
workbench "site" selection.

To summarize:
- The workbench site selection service can be seen as "centralized"
  service that registers selection providers and selection listeners.
- The editor part and the ouline are selection providers.
- The editor part, the outline and the property sheet are listener
  which all listen to each others indirectly.
- Hilarity ensues.

I tried to add enough javadoc in the classes to explain what's
going on, so please tell me if more is needed. (Editor part also
has a link to the one web page article that brings some sense to
this stuff. I recommend reading the web page first.)

Change-Id: Ief83f9fe2fc1cb5c0c1fa9ae174a58c8daf17ac4