Fix a possible deadlock in Eclipse/ADT.
The issue is that sometimes phones will stop responding to shell
commands launched through adb.

One of these commands is executed in DeviceMonitor#queryNewDeviceForInfo
However this is done from a loop that acquires a lock that is used in
other places, including building. This basically renders eclipse
useless as Eclipse will refuse to do anything until the build is
finished.

The fix is twofold:
First, move the call to queryNewDeviceForInfo outside of the
synchronized block.

This fixes the deadlock issue but this will not fix the problem
that the device monitor will still be stuck with a shell command
that never returns.

The second part of the fix is to add a timeout to shell command
execution. Additionnaly, this patch contains a lot of clean up
of the adb API in ddmlib, especially around exceptions thrown
during timeout.

Change-Id: Ice8ef787c825e0e7b535ff0bb939bf6f25e3d7e4